# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handler.SwaggerError:
    properties:
      code:
        example: VALIDATION_FAILED
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        example: Validation failed
        type: string
      requestId:
        example: req-abc123
        type: string
      timestamp:
        example: "2024-02-01T10:30:00Z"
        type: string
    type: object
  internal_handler.SwaggerErrorResponse:
    description: Standard error response format
    properties:
      error:
        $ref: '#/definitions/internal_handler.SwaggerError'
    type: object
...
  internal_handler.SwaggerJobError:
    description: Job error information
    properties:
      code:
        example: DOWNLOAD_FAILED
        type: string
      message:
        example: Failed to download video
        type: string
      retryable:
        example: true
        type: boolean
    type: object
  internal_handler.SwaggerJobResponse:
    description: Video extraction job status
    properties:
      completedAt:
        example: "2024-02-01T10:31:30Z"
        type: string
      createdAt:
        example: "2024-02-01T10:30:00Z"
        type: string
      error:
        $ref: '#/definitions/internal_handler.SwaggerJobError'
      estimatedSeconds:
        example: 45
        type: integer
      jobId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      message:
        example: Downloading video...
        type: string
      progress:
        example: 45
        type: integer
      recipe:
        $ref: '#/definitions/internal_handler.SwaggerRecipe'
      status:
        enum:
        - pending
        - downloading
        - processing
        - extracting
        - completed
        - failed
        - cancelled
        example: processing
        type: string
      statusUrl:
        example: /api/v1/jobs/550e8400-e29b-41d4-a716-446655440000
        type: string
      streamUrl:
        example: /api/v1/jobs/550e8400-e29b-41d4-a716-446655440000/stream
        type: string
    type: object
...
  internal_handler.SwaggerRecipe:
    description: Full recipe with ingredients and steps
    properties:
      cookTime:
        example: 20
        type: integer
      createdAt:
        example: "2024-02-01T10:30:00Z"
        type: string
      cuisine:
        example: Italian
        type: string
      description:
        example: Classic Italian pasta dish
        type: string
      difficulty:
        enum:
        - easy
        - medium
        - hard
        example: medium
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      ingredients:
        items:
          $ref: '#/definitions/internal_handler.SwaggerRecipeIngredient'
        type: array
      isFavorite:
        example: false
        type: boolean
      prepTime:
        example: 15
        type: integer
      servings:
        example: 4
        type: integer
      sourceType:
        enum:
        - manual
        - video
        - ai
        - photo
        - webpage
        - image
        example: manual
        type: string
      sourceUrl:
        example: https://youtube.com/watch?v=xyz
        type: string
      steps:
        items:
          $ref: '#/definitions/internal_handler.SwaggerRecipeStep'
        type: array
      syncVersion:
        example: 1
        type: integer
      tags:
        example:
        - italian
        - pasta
        - quick
        items:
          type: string
        type: array
      thumbnailUrl:
        example: https://example.com/thumb.jpg
        type: string
      title:
        example: Spaghetti Carbonara
        type: string
      updatedAt:
        example: "2024-02-01T10:30:00Z"
        type: string
      userId:
        example: 550e8400-e29b-41d4-a716-446655440099
        type: string
    type: object
  internal_handler.SwaggerRecipeIngredient:
    description: Recipe ingredient
    properties:
      category:
        enum:
        - dairy
        - produce
        - proteins
        - bakery
        - pantry
        - spices
        - condiments
        - beverages
        - snacks
        - frozen
        - household
        - other
        example: condiments
        type: string
      createdAt:
        example: "2024-02-01T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      isOptional:
        example: false
        type: boolean
      name:
        example: Olive oil
        type: string
      notes:
        example: extra virgin preferred
        type: string
      quantity:
        example: 2
        type: number
      recipeId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      sortOrder:
        example: 0
        type: integer
      unit:
        example: tablespoons
        type: string
      videoTimestamp:
        example: 120
        type: integer
    type: object
...
  internal_handler.SwaggerRecipeStep:
    description: Recipe step
    properties:
      createdAt:
        example: "2024-02-01T10:30:00Z"
        type: string
      durationSeconds:
        example: 120
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      instruction:
        example: Heat olive oil in a large pan over medium heat
        type: string
      recipeId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      stepNumber:
        example: 1
        type: integer
      technique:
        example: sauteing
        type: string
      temperature:
        example: 350F
        type: string
      videoTimestampEnd:
        example: 180
        type: integer
      videoTimestampStart:
        example: 60
        type: integer
    type: object
...
  internal_handler.SwaggerVideoExtractRequest:
    description: Video extraction request
    properties:
      detailLevel:
        enum:
        - quick
        - detailed
        example: detailed
        type: string
      language:
        enum:
        - en
        - fr
        - es
        - auto
        example: en
        type: string
      videoUrl:
        example: https://youtube.com/watch?v=abc123
        type: string
    required:
    - videoUrl
    type: object
host: api.dishflow.app
info:
  contact:
    email: support@dishflow.app
    name: DishFlow Support
  description: Recipe management API with AI-powered extraction, pantry tracking,
    and shopping lists
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://dishflow.app/terms
  title: DishFlow API
  version: "1.0"
paths:
...
  /jobs:
    get:
      description: Get list of video extraction jobs for the current user
      produces:
      - application/json
      responses:
        "200":
          description: List of jobs
          schema:
            items:
              $ref: '#/definitions/internal_handler.SwaggerJobResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List user's jobs
      tags:
      - Jobs
  /jobs/{jobID}:
    get:
      description: Get the current status of a video extraction job
      parameters:
      - description: Job UUID
        in: path
        name: jobID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Job status
          schema:
            $ref: '#/definitions/internal_handler.SwaggerJobResponse'
        "400":
          description: Invalid job ID
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "404":
          description: Job not found
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get job status
      tags:
      - Jobs
  /jobs/{jobID}/cancel:
    post:
      description: Cancel a running video extraction job
      parameters:
      - description: Job UUID
        in: path
        name: jobID
        required: true
        type: string
      responses:
        "204":
          description: Job cancelled
        "400":
          description: Invalid job ID
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "404":
          description: Job not found
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel a job
      tags:
      - Jobs
...
  /video/extract:
    post:
      consumes:
      - application/json
      description: Submit video URL for AI-powered recipe extraction (async job)
      parameters:
      - description: Video URL and options
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SwaggerVideoExtractRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Job created
          schema:
            $ref: '#/definitions/internal_handler.SwaggerJobResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "429":
          description: Rate limit exceeded (5/hour)
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Start video recipe extraction
      tags:
      - Jobs
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {your_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...